% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AG2ordered.R
\name{AG2ordered}
\alias{AG2ordered}
\title{Change (start,stop) or Anderson-Gill style data into Ordered failure time and risk Set}
\usage{
AG2ordered(data, id = "id", status = "status", start = "start", stop = "stop")
}
\arguments{
\item{data}{A data.frame}

\item{id}{Character name of id variable}

\item{status}{Character Name of status variable}

\item{start}{Character Name of start variable}

\item{stop}{Character Name of stop variable}
}
\value{
A data.frame
}
\description{
Change (start,stop) or Anderson-Gill style data into Ordered failure time and risk Set
}
\examples{
data(cancer,package="survival")
AG2ordered(bladder2[1:51,],status="event")
}
