% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AG2WLW.R
\name{AG2WLW}
\alias{AG2WLW}
\title{Change (start,stop) or Anderson-Gill style data into WLW style}
\usage{
AG2WLW(
  data,
  id = "id",
  status = "status",
  start = "start",
  stop = "stop",
  interval = "interval"
)
}
\arguments{
\item{data}{A data.frame}

\item{id}{Character name of id variable}

\item{status}{Character Name of status variable}

\item{start}{Character Name of start variable}

\item{stop}{Character Name of stop variable}

\item{interval}{Character Name of interval variable}
}
\value{
A data.frame
}
\description{
Change (start,stop) or Anderson-Gill style data into WLW style
}
\examples{
data(cancer,package="survival")
result=AG2WLW(bladder2,status="event",interval="enum")
identical(result,bladder)
}
