% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2CIC.R
\name{data2CIC}
\alias{data2CIC}
\title{Calculate cumulative incidence curves}
\usage{
data2CIC(df, time = "time", status = "status", cause = 1, add.CPC = TRUE)
}
\arguments{
\item{df}{A data frame}

\item{time}{character name of time variable}

\item{status}{character name of status variable}

\item{cause}{Numeric status for CIC calculation}

\item{add.CPC}{logical Whether or not add conditional probability curves(CPC)}
}
\value{
A data.frame
}
\description{
Calculate cumulative incidence curves
}
\examples{
time=c(0.7,3,4.9,6,6,6.9,10,10.8,17.1,20.3,1.5,2.8,3.8,4.7,7,10,10,11.2,
3.2,7.6,10,11,15,24.4)
status=c(rep(1,10),rep(2,8),rep(0,6))
df=data.frame(time,status)
data2CIC(df,cause=1,add.CPC=FALSE)
data2CIC(df,cause=NULL,add.CPC=FALSE)
data2CIC(df)
}
