% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hide.R
\name{hide}
\alias{hide}
\title{Hide some cells and draw outline in flextable}
\usage{
hide(ft, i, j, color = "red", width = 2)
}
\arguments{
\item{ft}{A flextable}

\item{i}{rows selected}

\item{j}{columns selected}

\item{color}{color of border}

\item{width}{width of border}
}
\value{
a flextable
}
\description{
Hide some cells and draw outline in flextable
}
\examples{
library(survival)
data(anderson,package="autoReg")
fit=survfit(Surv(time,status)~rx,data=anderson)
ft=howto(fit)
hide(ft,i=5:7,j=6:7)
}
