% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/howto.R
\name{howto}
\alias{howto}
\title{Explain how to calculate survival rate}
\usage{
howto(fit)
}
\arguments{
\item{fit}{An object of class survfit}
}
\value{
a flextable or NULL
}
\description{
Explain how to calculate survival rate
}
\examples{
library(survival)
data(anderson,package="autoReg")
fit=survfit(Surv(time,status)~rx,data=anderson)
howto(fit)
fit1=survfit(Surv(time,status)~1,data=anderson)
howto(fit1)
fit2=survfit(Surv(time,status)~1,data=anderson[anderson$rx==1,])
howto(fit2)
}
