% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/howto2.R
\name{howto2}
\alias{howto2}
\title{Explain how to perform log-rank test}
\usage{
howto2(fit)
}
\arguments{
\item{fit}{An object of class survfit}
}
\value{
A flextable or NULL
}
\description{
Explain how to perform log-rank test
}
\examples{
library(survival)
data(anderson,package="autoReg")
fit=survfit(Surv(time,status)~rx,data=anderson)
survdiff(Surv(time,status)~rx,data=anderson)
howto2(fit)
}
