% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WLW2AG.R
\name{WLW2AG}
\alias{WLW2AG}
\title{Change WLW style data into (start,stop) or Anderson-Gill style data}
\usage{
WLW2AG(
  data,
  id = "id",
  status = "status",
  start = "start",
  stop = "stop",
  interval = "interval"
)
}
\arguments{
\item{data}{A data.frame}

\item{id}{Character name of id variable}

\item{status}{Character Name of status variable}

\item{start}{Character Name of start variable}

\item{stop}{Character Name of stop variable}

\item{interval}{Character Name of interval variable}
}
\value{
A data.frame
}
\description{
Change WLW style data into (start,stop) or Anderson-Gill style data
}
\examples{
data(cancer,package="survival")
result=WLW2AG(bladder,status="event",interval="enum")
identical(result,bladder2)
}
